refactor(canvas): extract AnimatedThreadDock; fix resize-handle clip + a11y
Address review findings on the thread slide-in:
- Extract the duplicated AnimatePresence/width-slide wrapper (Activity + channel)
  into a shared AnimatedThreadDock.
- Apply overflow-hidden only while animating, so the settled panel no longer
  clips the ResizableSidebar resize handle (it overhangs the inner edge).
- Guard the mention-pulse and jump-pill transition with prefers-reduced-motion.
- Give all highlighted thread rows rounded-none for a consistent full-bleed pulse.
